Returned original function http = query history
