+++ .gitignore
@@ -1,5 +1,6 @@
 # Serverless
 .serverless
+.keyfiles
 .env
 tmp
 .coveralls.yml
