+++ keyfiles/Finance-asset-Serverless-service-account.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "platinum-form-233310",
+  "private_key_id": "2081133ca69a5afff2b1d94227cf7ab90e322bb6",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQC3KIXXzGukV226\nMQXfIFvUHqUUz2VzWt3X6DGG2TSItUUr4MeO9F4DtNYrwz8CtjUKAnvJV69ufBl/\nvIiBZRTObt4cvTSRnfl3R2HGcKUin//MjPtmt1CwyAIIufKR5uq70oPKBjk65YOr\necmUCmaaODFf/uBZMP61Qq0Adt0+cW4OguE+JCRQWgT1CLiLn024rv2YUQ6WThsV\n4rMIuAwqnUPAaTMl60JYYzI+R4uGVG2CaM5XMBw8yByOZZvOOTzpnMhGQZXxb9eP\nQC1JZiS2FcOcOjGvMN3aKe3Btm683fi84tialZfu8zicQcfUvHpgMuZlkMxTE+5H\n4ux7NL3TAgMBAAECggEAT/m1S6/alrWIV6M9TZr7DendqnMO+OpF8RdVvOyN+F2y\nmQC7EoxYmuS1GgHUuzHlgfvooleT5jnxBmpD7sA88BDQEpt5X1YsDm3QX05jhPV5\nRA+RXN4RsUd8P+Q164h5mc7Htgi2p6V9M3xnzSkneRKtWJ3Zx4hRx8MX+Jpua+tr\nvsA6ziHI6IK0CCgXOAKOulZqMs6ADq2PgG2eeaWJK1fee3lfCPvUmZjERLr6ahs0\nOUMH5eBZP1bcCDHJ3xSX1Dppf4MiWnK2i65HvuTpF5WF8HLE3XlvCxPVFEdwAfVy\nrS7yk8AFOV0tSGBQkfNRG26GsjfJ8SK39GPoBiYyUQKBgQD4dAaDhx4BWT+Adz7o\na0EKW5WdFP1elyJQpV8/ajz4dIaDrWn4qKKDF16aYQNcPBzMLw/slflzjPSt2lwU\nPRW8onKVLzXhtwOYIRIQ4HxKO1tIvxg+T0v5YysmEZcF6hkUpka/oYZP3uePTQNN\ndcmze01jRNpalPuLZEkuUC6w4wKBgQC8uMNtXxk3wd/ixNOXivS8Pc9Cj2gTjc/r\nXEgBz8N5JliTYKKbbacFx4+RnHXK+E/qUeHcH8c9w44bgGYQsxtFjMHZJhP+W4o4\nlVDCUZXvyLoCk+CaRp1lkTrr/UkJXf1o7dnnAP1GDiwaMViVJMB7jdoS4az0rN0x\n7Drog08CUQKBgQCNxf91jGOl11yyHNmL3LAHjg1IzpJqKInPRR6d3CyghjW/6/O9\njdT2Qx6br2IwfqjcJrrKgSYQNs3kaucCMX023+fMk10GyN1HMQ/TC5ZA8a/8Sfc2\njx9GVr1ngsAQAu5+5wcWEMDI0OSH1CVfTShkv6aAT+jdqYgEmZxczipmiwKBgHFQ\nJtcmwId/+yO3tqxJHMVh/24hmQYUogDZecrMqru1UH29CiV041JPDl8JqJhkG+e/\nnnODCSiGoG7W23GvePJ7n3D4sgyJfiaXUNQ8FGDTDwy5KaS4PY1cdBp6DD55lsxH\n9AwFxg0oOUbI1j1GwePtukCYDDclC/MLuhsdrHThAoGAOd8/e1XYLX304N/U9aJX\n+Itgavnlwzhbce4Gw3EJpiOe/ShC51YAykkv8arvfzhlgqE9sseSYQl/pGqtP+ew\n86T2+2Tkbgt0Fg7LJRYZFoNckt1p9XCoiMCuiER0/EH6xJ7gn2GpD9z1Pbv8VWAE\n/TLFAuqiV2T3wGjfm0IUJqU=\n-----END PRIVATE KEY-----\n",
+  "client_email": "financeasset@platinum-form-233310.iam.gserviceaccount.com",
+  "client_id": "102935734657266766795",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/financeasset%40platinum-form-233310.iam.gserviceaccount.com"
+}
