+++ serverless.yml
@@ -10,7 +10,7 @@ provider:
   # https://serverless.com/framework/docs/providers/google/guide/credentials/
   #
   # the path to the credentials file needs to be absolute
-  credentials: ../../gcloud/keyfile.json
+  credentials: ./keyfiles/Finance-asset-Serverless-service-account.json
 
 plugins:
   - serverless-google-cloudfunctions
@@ -22,6 +22,7 @@ package:
     - node_modules/**
     - .gitignore
     - .git/**
+    - keyfiles/**
 
 functions:
   first:
